Extract TraktRatingCollection to own file
